single file component: move Milestone.ts to Milestone.vue
